requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
twilio==8.11.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyMySQL==1.1.0
cryptography==41.0.7
SQLAlchemy==2.0.23
python-dateutil==2.8.2
Werkzeug==3.0.1